% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbcoef}
\alias{blbcoef}
\title{compute the coefficients from fit
compute the coefficients from each model}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{the linear regression model for each subsample}
}
\description{
compute the coefficients from fit
compute the coefficients from each model
}
