% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{find the logistic regression with bag of little bootstraps
blbglm is used to fit logistic models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance}
\usage{
blbglm(formula, data, m = 10, B = 5000, family)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class)}

\item{data}{an data frame, list or environment containing the variables in the model.}

\item{m}{split data into m parts，default number is 10}

\item{B}{eachoots run B times,dedefault number is 5000}
}
\description{
find the logistic regression with bag of little bootstraps
blbglm is used to fit logistic models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance
}
