% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: linear regression model and logistic regression model}
\description{
There are some usual function to estimate linear regression model and logistic regression model.
}
\details{
Regression models with Little Bag of Bootstraps
}
