% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{find the linear regression with bag of little bootstraps
blblm is used to fit linear regression  models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{an data frame, list or environment containing the variables in the model.}

\item{m}{split data into m parts，default number is 10}

\item{B}{times of bootstrap}
}
\description{
find the linear regression with bag of little bootstraps
blblm is used to fit linear regression  models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance
}
