% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{coefficients for blb regression model}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{fitted model}

\item{...}{extra conditions}
}
\description{
coefficients for blb regression model
}
