% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the logistic regression estimates based on given number of repetitions}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{the sub_data from original data set}

\item{freqs}{the frequence from glm_each_boot function}

\item{family}{family in glm}
}
\description{
estimate the logistic regression estimates based on given number of repetitions
}
