% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute the GLM estimates for a blb dataset}
\usage{
glm_each_boot(formula, data, n, family)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{n}{number of random vectors to draw}

\item{family}{family in glm}
}
\description{
compute the GLM estimates for a blb dataset
}
