% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{compute the estimates for GLM}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{n}{data}

\item{B}{times bootstrap}

\item{family}{family in glm}
}
\description{
compute the estimates for GLM
}
