% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class)}

\item{data}{the sub_data from original data set}

\item{freqs}{the frequence from lm_each_boot function}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
