% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset.}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class)}

\item{data}{the sub_data from original data set}

\item{n}{the size of original data set}
}
\description{
compute the regression estimates for a blb dataset.
}
