% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates
compute the estimates for each subsample and fit linear regression
model for each subsmple B times}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{the sub_data from original data set}

\item{n}{the size of original data set}

\item{B}{the subsample repeat B times}
}
\description{
compute the estimates
compute the estimates for each subsample and fit linear regression
model for each subsmple B times
}
