% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blbglm}
\alias{par_blbglm}
\title{find the logistics regression with bag of little bootstraps useing future_map function
par_blblm is used to fit linear regression  models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance
It can speed up the calculation}
\usage{
par_blbglm(formula, data, m = 10, B = 5000, family)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class)}

\item{data}{an data frame, list or environment containing the variables in the model.}

\item{m}{split data into m parts，default number is 10}

\item{B}{times of bootstrap}
}
\description{
find the logistics regression with bag of little bootstraps useing future_map function
par_blblm is used to fit linear regression  models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance
It can speed up the calculation
}
