% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{an data frame containing the variables in the function.}

\item{m}{split data into m parts，default number is 10}
}
\description{
split data into m parts of approximated equal sizes
}
